kfl/requirements.txt
amqp==5.3.1
asgiref==3.6.0
asttokens==2.4.1
attrs==22.2.0
billiard==4.2.1
black==23.7.0
celery==5.4.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.18
django-celery-beat==2.7.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.1
django-jazzmin==3.0.1
django-phonenumber-field==7.0.2
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-xml==2.0.0
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.4.1
exceptiongroup==1.1.3
executing==2.0.1
Faker==35.2.0
gunicorn==20.1.0
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
ipython==8.17.2
isort==5.12.0
jedi==0.19.1
jsonschema==4.17.3
kombu==5.4.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
PyMuPDF==1.25.4
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
phonenumbers==8.13.8
pillow==11.1.0
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
prompt-toolkit==3.0.41
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.2
redis==5.2.1
requests==2.31.0
setuptools==75.8.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.3
tomli==2.0.1
traitlets==5.13.0
typing_extensions==4.7.1
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.10
zipp==3.15.0
